Fuzzybool uses tnorm to prevent altering animation easings

+++ src/chart/rendering/drawitem.cpp
@@ -182,7 +182,7 @@ void drawItem::draw(
 		{
 			painter.drawStraightLine(
 				drawItem.getLine(), drawItem.lineWidth,
-				colors.second, colors.second * drawItem.connected);
+				colors.second, colors.second * (double)drawItem.connected);
 		}
 	}
 	else
